DBC-1020-100 ビジネス_福祉用具購入費支給申請 実装単体.福祉用具購入費支給申請一括審査・決定

+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/ShichosonTokubetsuKyufuJigyoshaIdentifier.java
@@ -6,10 +6,9 @@
 package jp.co.ndensan.reams.db.dbc.business.core.basic;
 
 import java.io.Serializable;
-import lombok.Value;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.JigyoshaNo;
 import jp.co.ndensan.reams.db.dbx.definition.core.valueobject.domain.ServiceCode;
-import jp.co.ndensan.reams.uz.uza.math.Decimal;
+import lombok.Value;
 
 /**
  * 市町村特別給付サービス事業者の識別子です。
@@ -19,7 +18,7 @@ public class ShichosonTokubetsuKyufuJigyoshaIdentifier implements Serializable {
 
     private final JigyoshaNo 市町村特別給付用事業者番号;
     private final ServiceCode 市町村特別給付用サービスコード;
-    private final Decimal 履歴番号;
+    private final int 履歴番号;
 
     /**
      * コンストラクタです。
@@ -30,7 +29,7 @@ public class ShichosonTokubetsuKyufuJigyoshaIdentifier implements Serializable {
      */
     public ShichosonTokubetsuKyufuJigyoshaIdentifier(JigyoshaNo 市町村特別給付用事業者番号,
             ServiceCode 市町村特別給付用サービスコード,
-            Decimal 履歴番号) {
+            int 履歴番号) {
         this.市町村特別給付用事業者番号 = 市町村特別給付用事業者番号;
         this.市町村特別給付用サービスコード = 市町村特別給付用サービスコード;
         this.履歴番号 = 履歴番号;
